+++ src/main/java/scripting/event/EventInstanceManager.java
@@ -892,7 +892,7 @@ public class EventInstanceManager {
             
             if (ServerConstants.JAVA_8) {
                 for (Object d: list) {
-                    intList.add(((Integer) d).intValue());
+                    intList.add((Integer) d);
                 }
             } else {
                 for (Object d: list) {
